+++ fast_llm/functional/triton/entropy_loss.py
